refactor map loading logic
