+++ workspaces/maploading/assets/level1/level1.json
@@ -1,117 +1,48 @@
 {
   "collisionAreas": [],
   "entities": {
-    "77eo8n": {
-      "attributes": {
-        "oneWay": false
-      },
+    "2gkQ3n": {
       "boundary": {
-        "height": 158,
-        "startX": 304,
-        "startY": 1082,
-        "width": 272
+        "height": 23,
+        "startX": 127,
+        "startY": 1641,
+        "width": 321
       },
       "collision": {
-        "height": 80,
-        "startX": 363,
-        "startY": 1124,
-        "width": 155
+        "height": 23,
+        "startX": 127,
+        "startY": 1641,
+        "width": 321
       },
       "type": "platform"
     },
-    "BtSAoe": {
-      "attributes": {
-        "oneWay": false
-      },
+    "3nzzz6": {
       "boundary": {
-        "height": 26,
-        "startX": 191,
-        "startY": 870,
-        "width": 257
+        "height": 25,
+        "startX": 448,
+        "startY": 1191,
+        "width": 320
       },
       "collision": {
-        "height": 26,
-        "startX": 191,
-        "startY": 870,
-        "width": 257
+        "height": 25,
+        "startX": 448,
+        "startY": 1191,
+        "width": 320
       },
       "type": "platform"
     },
-    "F4bOc9": {
-      "attributes": {
-        "oneWay": false
-      },
+    "uQjBCL": {
       "boundary": {
-        "height": 116,
-        "startX": 52,
-        "startY": 1914,
-        "width": 298
-      },
-      "type": "platform"
-    },
-    "TdSXO7": {
-      "boundary": {
-        "height": 131,
-        "startX": 314,
-        "startY": 51,
-        "width": 190
+        "height": 23,
+        "startX": 960,
+        "startY": 1897,
+        "width": 256
       },
       "collision": {
-        "height": 131,
-        "startX": 314,
-        "startY": 51,
-        "width": 190
-      },
-      "type": "platform"
-    },
-    "WdSXfK": {
-      "attributes": {
-        "oneWay": false
-      },
-      "boundary": {
-        "height": 236,
-        "startX": 773,
-        "startY": 1272,
-        "width": 389
-      },
-      "collision": {
-        "height": 153,
-        "startX": 865,
-        "startY": 1317,
-        "width": 275
-      },
-      "type": "platform"
-    },
-    "ad9nQh": {
-      "attributes": {
-        "oneWay": false
-      },
-      "boundary": {
-        "height": 157,
-        "startX": 1104,
-        "startY": 931,
-        "width": 108
-      },
-      "collision": {
-        "height": 120,
-        "startX": 1121,
-        "startY": 950,
-        "width": 85
-      },
-      "type": "platform"
-    },
-    "qVR6JO": {
-      "boundary": {
-        "height": 125,
-        "startX": 732,
-        "startY": 843,
-        "width": 212
-      },
-      "collision": {
-        "height": 125,
-        "startX": 732,
-        "startY": 843,
-        "width": 212
+        "height": 23,
+        "startX": 960,
+        "startY": 1897,
+        "width": 256
       },
       "type": "platform"
     }
