+++ workspaces/maploading/src/util/map/bundlebuilder.rs
@@ -1,9 +1,9 @@
 use bevy::prelude::*;
 
-use crate::{Screen, CameraController};
-
 #[derive(Component)]
 pub struct FullscreenSprite;
+
+use super::data::CameraController;
 
 pub fn full_image(
     map_dimentions: &(u32, u32),
@@ -17,18 +17,17 @@ pub fn full_image(
     )
 }
 
-pub fn camera_start(screen: Screen) -> impl Bundle {
+pub fn camera_start(screen: (u32, u32)) -> impl Bundle {
     (
         Camera2d,
         Transform {
             translation: Vec3::new(
-                screen.x as f32 / 2.0,
-                screen.y as f32 / 2.0,
+                screen.0 as f32 / 2.0,
+                screen.1 as f32 / 2.0,
                 0.0, // keep positive z so it's above everything
             ),
             scale: Vec3::splat(1.0),
-            ..Default::default()
-        },
+            ..Default::default() },
         CameraController,
     )
 }
