+++ workspaces/maploading/src/util/map/data.rs
@@ -1,0 +1,56 @@
+// src/util/map/dat.rs
+use bevy::math::primitives::Rectangle;
+use bevy::prelude::*;
+use serde::Deserialize;
+use std::collections::HashMap;
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct Metadata {
+    pub tile_size_px: u32,
+    pub rows: u32,
+    pub cols: u32,
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct Boundary {
+    pub start_x: f32,
+    pub start_y: f32,
+    pub width: f32,
+    pub height: f32,
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct EntityData {
+    pub boundary: Boundary,
+    #[serde(rename = "type")]
+    pub kind: String,
+}
+
+#[derive(Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct LayerImages {
+    pub tile_fg: String,
+    pub entity: String,
+}
+
+#[derive(Deserialize, Debug, Resource)]
+#[serde(rename_all = "camelCase")]
+pub struct MapFile {
+    pub metadata: Metadata,
+    pub layer_images: LayerImages,
+    pub collision_areas: Vec<Rectangle>,
+    pub entities: HashMap<String, EntityData>,
+}
+
+// Bevy data structs
+#[derive(Resource)]
+pub struct MapTextureHandles {
+    pub tile_fg: Handle<Image>,
+    pub entity: Handle<Image>,
+}
+
+#[derive(Component)]
+pub struct CameraController;
