+++ workspaces/maploading/src/util/map/mod.rs
@@ -1,0 +1,9 @@
+pub mod spawn;
+pub mod loader;
+pub mod data;
+pub mod bundlebuilder;
+
+pub const MAP_NAME: &str = "level1";
+pub const SCREEN_WIDTH: u32 = 1280;
+pub const SCREEN_HEIGHT: u32 = 720;
+
