+++ workspaces/maploading/src/util/mod.rs
@@ -1,3 +1,2 @@
-pub mod maploading;
-pub mod setup;
+pub mod map;
 pub mod dev_mode;
